Added more to Score class. Played around with game class.
Score is now centered at the bottom of any screen and updates while
remaining entered. Added a method to return the current dimensions of
the score text, may be useful in ensuring the robot doesn’t overlap it.
Also added a bit to the game class to test if it is registering a hit
on the robot.